+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Add `simplify` flag to ONNX export task to simplify the model after export.
+
 ### Deprecated
 
 ### Removed